requirements.txt
flask>=3.0
scikit-learn>=1.4
numpy>=1.26
pdfplumber>=0.11
gunicorn>=21.2
psycopg2-binary>=2.9
joblib>=1.3
